refactor: build menu structure in regular code flow instead of separate script
Before, the menu structure was build via a separate script which
exported to a json file. That file was read and parsed during code
execturion (build time). However, that meant having the file written to
disk, which should be gitignored.

Having it part of the regular code flow, there is no need for a
dedicated json file.